requirements.txt
discord.py
Pillow
requests
